requirements.txt
pinboard==2.1.4
mercury-parserpy==0.4
urlextract==0.8.3
